--- hibernate-core/src/main/java/org/hibernate/metamodel/mapping/IEmbeddableMappingType.java
@@ -1,53 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
- */
-package org.hibernate.metamodel.mapping;
-
-import java.util.List;
-
-import org.hibernate.mapping.IndexedConsumer;
-import org.hibernate.metamodel.mapping.internal.EmbeddedAttributeMapping;
-import org.hibernate.metamodel.mapping.internal.MappingModelCreationProcess;
-import org.hibernate.metamodel.spi.EmbeddableRepresentationStrategy;
-import org.hibernate.sql.ast.tree.from.TableGroupProducer;
-
-/**
- * @author Steve Ebersole
- */
-public interface IEmbeddableMappingType extends ManagedMappingType, SelectableMappings {
-	IEmbeddableMappingType createInverseMappingType(
-			EmbeddedAttributeMapping valueMapping,
-			TableGroupProducer declaringTableGroupProducer,
-			SelectableMappings selectableMappings,
-			MappingModelCreationProcess creationProcess);
-
-	EmbeddableValuedModelPart getEmbeddedValueMapping();
-
-	EmbeddableRepresentationStrategy getRepresentationStrategy();
-
-	Object[] getPropertyValues(Object compositeInstance);
-
-	void setPropertyValues(Object compositeInstance, Object[] resolvedValues);
-
-	boolean isCreateEmptyCompositesEnabled();
-
-	@Override
-	default int forEachSelectable(SelectableConsumer consumer) {
-		return ManagedMappingType.super.forEachSelectable( consumer );
-	}
-
-	@Override
-	int forEachSelectable(int offset, SelectableConsumer consumer);
-
-	@Override
-	int getJdbcTypeCount();
-
-	@Override
-	List<JdbcMapping> getJdbcMappings();
-
-	@Override
-	int forEachJdbcType(int offset, IndexedConsumer<JdbcMapping> action);
-}
