+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/embeddable/AbstractEmbeddableInitializer.java
@@ -17,7 +17,7 @@ import org.hibernate.metamodel.mapping.CompositeIdentifierMapping;
 import org.hibernate.metamodel.mapping.EmbeddableValuedModelPart;
 import org.hibernate.metamodel.mapping.EntityIdentifierMapping;
 import org.hibernate.metamodel.mapping.ForeignKeyDescriptor;
-import org.hibernate.metamodel.mapping.IEmbeddableMappingType;
+import org.hibernate.metamodel.mapping.EmbeddableMappingType;
 import org.hibernate.metamodel.mapping.StateArrayContributorMapping;
 import org.hibernate.metamodel.mapping.internal.ToOneAttributeMapping;
 import org.hibernate.metamodel.spi.EmbeddableRepresentationStrategy;
@@ -62,7 +62,7 @@ public abstract class AbstractEmbeddableInitializer extends AbstractFetchParentA
 		this.embeddedModelPartDescriptor = resultDescriptor.getReferencedMappingContainer();
 		this.fetchParentAccess = fetchParentAccess;
 
-		final IEmbeddableMappingType embeddableTypeDescriptor = embeddedModelPartDescriptor.getEmbeddableTypeDescriptor();
+		final EmbeddableMappingType embeddableTypeDescriptor = embeddedModelPartDescriptor.getEmbeddableTypeDescriptor();
 		final int numOfAttrs = embeddableTypeDescriptor.getNumberOfAttributeMappings();
 		this.resolvedValues = new Object[ numOfAttrs ];
 		this.assemblerMap = new IdentityHashMap<>( numOfAttrs );
@@ -142,7 +142,7 @@ public abstract class AbstractEmbeddableInitializer extends AbstractFetchParentA
 		// Special handling for non-aggregated attributes which use the actual entity instance as container,
 		// which we access through the fetch parent access.
 		// If this model part is an identifier, we must construct the instance as this is called during resolveKey
-		final IEmbeddableMappingType embeddableTypeDescriptor = embeddedModelPartDescriptor.getEmbeddableTypeDescriptor();
+		final EmbeddableMappingType embeddableTypeDescriptor = embeddedModelPartDescriptor.getEmbeddableTypeDescriptor();
 		final EmbeddableRepresentationStrategy representationStrategy;
 		if ( embeddedModelPartDescriptor instanceof CompositeIdentifierMapping ) {
 			representationStrategy = ( (CompositeIdentifierMapping) embeddedModelPartDescriptor ).getMappedIdEmbeddableTypeDescriptor()
@@ -257,7 +257,7 @@ public abstract class AbstractEmbeddableInitializer extends AbstractFetchParentA
 	}
 
 	private void setPropertyValuesOnTarget(Object compositeInstance, SharedSessionContractImplementor session) {
-		final IEmbeddableMappingType embeddableTypeDescriptor;
+		final EmbeddableMappingType embeddableTypeDescriptor;
 		if ( embeddedModelPartDescriptor instanceof CompositeIdentifierMapping ) {
 			final CompositeIdentifierMapping compositeIdentifierMapping = (CompositeIdentifierMapping) this.embeddedModelPartDescriptor;
 			embeddableTypeDescriptor = compositeIdentifierMapping.getMappedIdEmbeddableTypeDescriptor();
