+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/embeddable/EmbeddableResultGraphNode.java
@@ -7,7 +7,7 @@
 package org.hibernate.sql.results.graph.embeddable;
 
 import org.hibernate.metamodel.mapping.EmbeddableValuedModelPart;
-import org.hibernate.metamodel.mapping.IEmbeddableMappingType;
+import org.hibernate.metamodel.mapping.EmbeddableMappingType;
 import org.hibernate.query.NavigablePath;
 import org.hibernate.sql.results.graph.DomainResultGraphNode;
 import org.hibernate.sql.results.graph.FetchParent;
@@ -25,5 +25,5 @@ public interface EmbeddableResultGraphNode extends DomainResultGraphNode, FetchP
 	EmbeddableValuedModelPart getReferencedMappingContainer();
 
 	@Override
-	IEmbeddableMappingType getReferencedMappingType();
+	EmbeddableMappingType getReferencedMappingType();
 }
