+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/embeddable/internal/EmbeddableFetchImpl.java
@@ -8,7 +8,7 @@ package org.hibernate.sql.results.graph.embeddable.internal;
 
 import org.hibernate.engine.FetchTiming;
 import org.hibernate.metamodel.mapping.EmbeddableValuedModelPart;
-import org.hibernate.metamodel.mapping.IEmbeddableMappingType;
+import org.hibernate.metamodel.mapping.EmbeddableMappingType;
 import org.hibernate.query.NavigablePath;
 import org.hibernate.query.results.ResultsHelper;
 import org.hibernate.sql.ast.SqlAstJoinType;
@@ -93,8 +93,8 @@ public class EmbeddableFetchImpl extends AbstractFetchParent implements Embeddab
 	}
 
 	@Override
-	public IEmbeddableMappingType getFetchContainer() {
-		return (IEmbeddableMappingType) super.getFetchContainer();
+	public EmbeddableMappingType getFetchContainer() {
+		return (EmbeddableMappingType) super.getFetchContainer();
 	}
 
 	@Override
@@ -133,7 +133,7 @@ public class EmbeddableFetchImpl extends AbstractFetchParent implements Embeddab
 	}
 
 	@Override
-	public IEmbeddableMappingType getReferencedMappingType() {
+	public EmbeddableMappingType getReferencedMappingType() {
 		return getFetchContainer();
 	}
